feat: merge protection and recovery.
logion-network/logion-internal#1309
