+++ src/wallet-user/UserPaths.tsx
@@ -13,8 +13,6 @@ export const TRUST_PROTECTION_RELATIVE_PATH = '/protection';
 export const TRUST_PROTECTION_PATH = USER_PATH + TRUST_PROTECTION_RELATIVE_PATH;
 export const SETTINGS_RELATIVE_PATH = '/settings';
 export const SETTINGS_PATH = USER_PATH + SETTINGS_RELATIVE_PATH;
-export const RECOVERY_RELATIVE_PATH = '/recovery';
-export const RECOVERY_PATH = USER_PATH + RECOVERY_RELATIVE_PATH;
 export const WALLET_RELATIVE_PATH = '/wallet';
 export const VAULT_RELATIVE_PATH = '/vault';
 export const ISSUER_RELATIVE_PATH = '/verified-issuer';
